Move internal/otlptext to exporter/loggingexporter/internal, only usage (#5483)
